+++ src/main/java/paibridge/apiheartee/counsel/entity/CounselReportDT.java
@@ -19,7 +19,7 @@ public class CounselReportDT extends CounselReport {
 
     @Builder
     public CounselReportDT(CounselRequest counselRequest, String summary,
-        String solution, Double willingness, Double selfOpenness, Double voiceOver, Double positiveLanguage, Double frequency, String explanation) {
+        String solution, Integer willingness, Integer selfOpenness, Integer voiceOver, Integer positiveLanguage, Integer frequency, String explanation) {
         super(counselRequest, summary, solution, willingness, selfOpenness, voiceOver, positiveLanguage, frequency, explanation);
 //        this.statDT = statDT;
     }
